requirements.txt
Flask==3.1.0
tensorflow-cpu==2.19.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Werkzeug==3.1.3
Pillow==11.1.0
numpy==2.1.3
gunicorn==20.1.0
